feat: CheckProfileGUID replace std::find with std::binary_search
The way of checking if the provided profile GUID exists in the vector of supported
GUIDs was modified. Instead of using std::find(), std::binary_search is used now.
This change assumes that the vector is sorted, which
might be the case if NvEncGetEncodeProfileGUIDs returns the GUIDs in sorted order.

Using std::binary_search can be more efficient than std::find for large datasets,
as it has a complexity of O(log n) instead of O(n).In summary, the changes optimize
the checking of whether the provided profile GUID exists in the vector of supported
GUIDs by leveraging the assumed sorted order of the GUIDs and using a binary search
instead of a linear search. This can potentially improve the performance of the code
for large datasets.