feat: ExecuteRPC replace switch statements
The switch statements have been replaced with simpler assignments,
which makes the code shorter and easier to read. The new assignments use the
ternary operator ? : to set the value of each variable based on the result
of the corresponding function call in the request object. For example,
access->enable is set to true if request.enable() returns SWITCH_ENABLE,
and false otherwise.

This makes the code more concise and more straightforward to understand.